ajout ope and/or/xor

+++ Corewar/srcs/corewar/operation_1.c
@@ -6,7 +6,7 @@
 /*   By: cpoulet <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/03/28 16:30:42 by cpoulet           #+#    #+#             */
-/*   Updated: 2017/04/03 16:59:17 by cpoulet          ###   ########.fr       */
+/*   Updated: 2017/04/04 12:08:11 by cpoulet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ void	op_ld(t_vm *v, t_list *process)
 			{
 				shift = reverse_bytes(&ARENA(PC + 2), 2);
 				shift %= IDX_MOD;
-				shift = reverse_bytes(&ARENA((unsigned int)(PC + shift)), 4);
+				shift = reverse_bytes(&ARENA((PC + shift)), 4);
 			}
 			REG[ARENA(PC + 6 - ((B_OCT & 0x60) >> 5)) - 1] = shift;
 			printf("val_saved = %x\n", shift);
@@ -49,8 +49,8 @@ void	op_ld(t_vm *v, t_list *process)
 
 void	op_st(t_vm *v, t_list *process)
 {
-	int	shift;
-	int	val;
+	int				shift;
+	unsigned int	val;
 
 	if ((B_OCT == 0x50 || B_OCT == 0x70) && ARENA(PC + 2) >= 1 && ARENA(PC + 2) <= 16)
 	{
@@ -63,7 +63,7 @@ void	op_st(t_vm *v, t_list *process)
 			val = REG[ARENA(PC + 2) - 1];
 			while (val)
 			{
-				ARENA((unsigned int)(PC + shift-- + 3)) = val & 0xFF;
+				ARENA((PC + shift-- + 3)) = val & 0xFF;
 				val >>= 8;
 			}
 		}
@@ -75,6 +75,6 @@ void	op_st(t_vm *v, t_list *process)
 void	op_aff(t_vm *v, t_list *process)
 {
 	if (B_OCT == 0x40 && ARENA(PC + 2) >= 1 && ARENA(PC + 2) <= 16)
-		printf("aff = |%d|\n", (u_char)REG[ARENA(PC + 2) - 1]);
+		printf("aff = |%d|\n", (u_char)REG[ARENA(PC + 2) - 1]); // a mettre au bon format output
 	octal_shift(process, B_OCT, 4, 1);
 }
